Add files via upload
Questions are given in the project and then solved with the help of Python. It is a project of Data Analysis with Python or you can say, Data Science with Python.

The commands that we used in this project :
head() - It shows the first N rows in the data (by default, N=5).
shape - It shows the total no. of rows and no. of columns of the dataframe
index - This attribute provides the index of the dataframe
columns - It shows the name of each column
dtypes - It shows the data-type of each column
unique() - In a column, it shows all the unique values. It can be applied on a single column only, not on the whole dataframe.
nunique() - It shows the total no. of unique values in each column. It can be applied on a single column as well as on the whole dataframe.
count - It shows the total no. of non-null values in each column. It can be applied on a single column as well as on the whole dataframe.
value_counts - In a column, it shows all the unique values with their count. It can be applied on a single column only.
info() - Provides basic information about the dataframe.